Gmaps.cs: Changes regarding distance calculations
Because of an unknown (for the time being) null exception return, while GMaps try to create GDirections between 2 points, there was some kind of "Brute force" method to calculate every distance from and to every point.

Also, because this might take a little bit of time, there was a ProgressBar added to show the current state of calculations.